Added default color compatibility mode and implemented both examples.
