--- CompositeColorVideo/luni.h
@@ -1,307 +1,0 @@
-namespace luni 
-{
-const unsigned int xres = 80;
-const unsigned int yres = 60;
-const unsigned char pixels[] = {
-35, 36, 36, 36, 37, 37, 37, 37, 38, 38, 38, 38, 39, 39, 39, 39, 
-39, 39, 39, 39, 39, 39, 40, 40, 40, 40, 39, 39, 39, 39, 39, 39, 
-39, 38, 38, 38, 38, 38, 38, 37, 37, 37, 37, 36, 36, 36, 35, 35, 
-35, 34, 34, 33, 33, 33, 32, 32, 32, 31, 31, 31, 30, 30, 26, 22, 
-22, 22, 22, 22, 22, 22, 22, 22, 22, 22, 21, 23, 25, 25, 25, 26, 
-35, 35, 35, 36, 36, 36, 37, 37, 37, 38, 38, 38, 38, 39, 39, 39, 
-39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 
-39, 39, 39, 38, 38, 38, 38, 38, 37, 37, 37, 37, 36, 36, 36, 36, 
-36, 35, 35, 34, 34, 34, 34, 33, 33, 33, 32, 32, 32, 31, 26, 22, 
-24, 24, 25, 24, 24, 24, 24, 23, 22, 20, 20, 25, 25, 26, 26, 26, 
-35, 35, 36, 36, 36, 37, 37, 37, 38, 38, 38, 39, 39, 39, 39, 39, 
-39, 39, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 39, 39, 
-39, 39, 39, 39, 39, 38, 38, 38, 38, 37, 37, 37, 37, 37, 36, 36, 
-36, 36, 35, 35, 35, 34, 34, 34, 33, 33, 33, 32, 32, 32, 27, 24, 
-27, 29, 29, 29, 29, 28, 28, 27, 26, 21, 19, 25, 26, 26, 26, 26, 
-36, 37, 37, 37, 38, 38, 38, 39, 39, 39, 39, 40, 40, 40, 40, 40, 
-40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 
-40, 40, 40, 39, 39, 39, 39, 39, 38, 38, 38, 37, 37, 37, 37, 37, 
-37, 36, 36, 36, 35, 35, 35, 35, 34, 34, 33, 33, 33, 32, 28, 24, 
-28, 29, 30, 29, 29, 29, 28, 28, 26, 22, 19, 25, 26, 26, 26, 26, 
-37, 37, 37, 38, 38, 39, 39, 39, 39, 40, 40, 40, 40, 40, 40, 40, 
-41, 41, 41, 41, 41, 41, 41, 41, 41, 41, 41, 41, 41, 41, 40, 40, 
-40, 40, 40, 40, 40, 39, 37, 34, 29, 31, 35, 37, 37, 36, 36, 37, 
-37, 37, 36, 36, 36, 35, 35, 35, 34, 34, 34, 33, 33, 33, 28, 25, 
-28, 30, 30, 30, 29, 29, 29, 28, 26, 22, 19, 26, 26, 26, 26, 25, 
-36, 37, 37, 38, 38, 38, 39, 39, 39, 39, 40, 40, 40, 40, 40, 40, 
-40, 40, 40, 40, 41, 41, 41, 41, 40, 40, 40, 40, 40, 40, 40, 40, 
-40, 40, 40, 39, 37, 33, 24, 17, 16, 14, 17, 22, 29, 28, 31, 36, 
-37, 36, 36, 36, 36, 35, 35, 35, 34, 34, 33, 33, 33, 33, 29, 25, 
-28, 30, 30, 30, 30, 29, 29, 29, 26, 22, 20, 25, 26, 26, 26, 25, 
-36, 36, 37, 37, 38, 38, 38, 39, 39, 39, 39, 39, 39, 40, 40, 40, 
-40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 
-40, 39, 39, 34, 23, 15, 13, 11, 11, 10, 10, 10, 14, 22, 28, 32, 
-35, 36, 36, 35, 35, 35, 35, 34, 34, 33, 33, 33, 33, 32, 29, 25, 
-28, 30, 30, 30, 30, 29, 29, 28, 26, 21, 20, 25, 26, 26, 25, 25, 
-36, 36, 36, 37, 37, 37, 38, 38, 38, 38, 39, 39, 39, 39, 39, 39, 
-39, 39, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 39, 39, 39, 39, 
-39, 38, 27, 14, 9, 6, 6, 9, 10, 11, 9, 10, 12, 9, 8, 12, 
-16, 32, 35, 35, 35, 34, 34, 34, 33, 33, 33, 32, 32, 32, 28, 25, 
-28, 30, 30, 30, 30, 29, 29, 28, 26, 22, 21, 25, 26, 25, 25, 24, 
-33, 34, 34, 34, 35, 36, 36, 36, 37, 37, 37, 38, 38, 38, 38, 38, 
-39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 
-37, 23, 11, 8, 8, 8, 6, 5, 9, 10, 10, 11, 10, 8, 8, 5, 
-5, 10, 26, 34, 34, 34, 34, 33, 33, 33, 32, 32, 32, 31, 28, 25, 
-28, 30, 30, 30, 29, 29, 28, 28, 25, 22, 21, 24, 25, 25, 25, 24, 
-13, 13, 13, 13, 13, 13, 14, 14, 15, 15, 15, 16, 16, 16, 17, 17, 
-17, 17, 18, 18, 18, 18, 18, 19, 19, 20, 20, 21, 21, 22, 22, 22, 
-20, 8, 4, 3, 5, 7, 8, 10, 10, 11, 11, 13, 11, 11, 6, 5, 
-5, 4, 8, 17, 21, 21, 22, 22, 22, 22, 22, 23, 24, 24, 24, 24, 
-28, 29, 30, 29, 29, 28, 26, 26, 25, 21, 21, 24, 25, 24, 25, 24, 
-9, 11, 12, 12, 13, 13, 13, 13, 14, 14, 14, 14, 14, 14, 14, 14, 
-14, 14, 15, 14, 14, 15, 15, 15, 15, 14, 15, 15, 15, 16, 17, 17, 
-10, 7, 6, 4, 7, 13, 14, 15, 15, 15, 15, 14, 14, 10, 6, 5, 
-7, 4, 4, 9, 15, 17, 17, 17, 17, 18, 19, 20, 20, 21, 23, 25, 
-26, 29, 29, 29, 28, 26, 25, 26, 24, 21, 21, 23, 24, 24, 24, 24, 
-16, 16, 16, 17, 17, 17, 17, 17, 17, 17, 18, 18, 18, 18, 18, 18, 
-18, 19, 19, 19, 19, 19, 19, 19, 19, 19, 19, 19, 19, 19, 17, 13, 
-5, 2, 5, 11, 16, 24, 26, 26, 23, 24, 25, 27, 24, 19, 14, 9, 
-8, 7, 3, 5, 13, 18, 18, 17, 22, 28, 24, 27, 28, 28, 28, 28, 
-27, 24, 27, 27, 25, 26, 24, 26, 23, 20, 21, 23, 24, 24, 24, 24, 
-13, 12, 13, 13, 14, 14, 15, 15, 15, 16, 16, 16, 16, 17, 17, 17, 
-18, 19, 20, 21, 21, 21, 21, 21, 21, 20, 20, 19, 20, 19, 13, 10, 
-2, 8, 22, 33, 36, 37, 37, 37, 36, 36, 36, 35, 33, 30, 25, 17, 
-11, 10, 3, 3, 10, 24, 25, 21, 25, 28, 25, 28, 29, 29, 29, 28, 
-28, 24, 21, 26, 22, 26, 25, 25, 20, 21, 21, 23, 24, 24, 24, 24, 
-6, 4, 4, 4, 4, 4, 4, 5, 5, 6, 6, 7, 9, 11, 14, 18, 
-24, 29, 32, 34, 35, 35, 35, 34, 33, 33, 33, 33, 33, 20, 12, 7, 
-6, 22, 35, 39, 40, 40, 40, 39, 39, 38, 36, 35, 33, 29, 25, 22, 
-16, 10, 4, 2, 4, 21, 27, 22, 25, 29, 25, 28, 29, 29, 29, 28, 
-28, 27, 17, 24, 23, 24, 26, 22, 22, 21, 22, 23, 23, 24, 24, 24, 
-11, 3, 4, 4, 4, 4, 4, 5, 5, 5, 6, 7, 11, 16, 23, 32, 
-43, 50, 52, 52, 52, 52, 52, 52, 52, 52, 52, 52, 50, 18, 11, 6, 
-15, 31, 39, 41, 40, 40, 39, 39, 39, 37, 36, 34, 31, 29, 26, 24, 
-17, 9, 4, 3, 2, 16, 25, 26, 26, 29, 25, 28, 29, 28, 27, 26, 
-25, 26, 24, 18, 23, 25, 24, 24, 22, 21, 22, 23, 23, 24, 24, 24, 
-8, 3, 4, 4, 4, 4, 4, 4, 5, 5, 6, 6, 7, 9, 12, 18, 
-27, 35, 38, 39, 38, 38, 37, 37, 36, 35, 35, 34, 28, 11, 8, 7, 
-23, 34, 39, 40, 40, 40, 38, 39, 38, 37, 35, 33, 32, 29, 27, 24, 
-15, 7, 4, 3, 2, 8, 24, 27, 26, 29, 24, 28, 27, 26, 25, 23, 
-22, 22, 21, 15, 23, 25, 25, 25, 22, 21, 22, 23, 23, 24, 24, 24, 
-4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 6, 6, 6, 6, 6, 7, 
-8, 9, 10, 10, 11, 11, 12, 12, 13, 13, 14, 14, 9, 10, 6, 13, 
-27, 36, 39, 40, 40, 39, 38, 39, 38, 36, 35, 32, 31, 29, 27, 24, 
-17, 10, 4, 3, 3, 4, 17, 27, 26, 29, 25, 28, 28, 27, 25, 22, 
-21, 24, 24, 18, 21, 24, 24, 25, 22, 21, 22, 23, 23, 23, 24, 24, 
-4, 4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 6, 6, 6, 7, 7, 
-8, 9, 9, 10, 11, 12, 12, 13, 13, 14, 14, 14, 7, 7, 6, 17, 
-30, 36, 39, 39, 40, 39, 39, 40, 38, 36, 34, 31, 29, 29, 28, 25, 
-20, 14, 5, 3, 2, 3, 10, 25, 26, 28, 25, 28, 28, 26, 26, 27, 
-27, 26, 25, 22, 21, 23, 22, 22, 22, 21, 22, 23, 23, 23, 23, 23, 
-4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 6, 6, 6, 7, 7, 7, 
-8, 8, 9, 9, 10, 10, 11, 11, 12, 12, 13, 11, 7, 5, 7, 21, 
-33, 37, 38, 39, 39, 38, 38, 38, 36, 34, 32, 30, 30, 30, 29, 25, 
-21, 17, 7, 3, 2, 3, 6, 19, 26, 28, 25, 28, 28, 28, 28, 28, 
-27, 26, 26, 23, 21, 24, 21, 22, 21, 21, 22, 23, 22, 23, 23, 23, 
-5, 5, 5, 5, 5, 5, 6, 6, 6, 7, 7, 7, 8, 8, 9, 9, 
-10, 10, 11, 11, 12, 12, 13, 14, 14, 14, 15, 10, 6, 4, 9, 26, 
-36, 37, 36, 34, 34, 34, 36, 34, 30, 27, 25, 26, 27, 28, 26, 24, 
-21, 19, 11, 3, 2, 2, 5, 10, 26, 28, 25, 28, 28, 28, 28, 27, 
-26, 26, 26, 22, 19, 23, 23, 20, 18, 18, 21, 23, 22, 22, 22, 22, 
-5, 5, 5, 5, 6, 6, 6, 7, 7, 8, 8, 8, 9, 10, 10, 11, 
-11, 12, 13, 14, 14, 15, 15, 16, 17, 17, 18, 12, 5, 4, 13, 28, 
-31, 30, 19, 18, 25, 32, 36, 28, 22, 13, 10, 11, 18, 19, 19, 20, 
-20, 19, 15, 5, 2, 2, 4, 5, 21, 28, 24, 28, 28, 28, 28, 27, 
-25, 23, 21, 22, 18, 21, 18, 20, 19, 19, 20, 21, 21, 22, 22, 22, 
-6, 5, 5, 6, 6, 6, 6, 7, 7, 8, 8, 9, 9, 10, 11, 11, 
-12, 12, 13, 14, 15, 15, 16, 17, 17, 18, 18, 12, 4, 4, 15, 24, 
-23, 21, 20, 16, 19, 28, 36, 25, 16, 14, 13, 17, 19, 17, 15, 15, 
-18, 19, 18, 7, 1, 1, 3, 5, 12, 27, 24, 28, 28, 28, 27, 24, 
-24, 20, 20, 25, 18, 25, 19, 19, 19, 20, 21, 22, 21, 20, 21, 22, 
-5, 5, 5, 5, 6, 6, 6, 7, 7, 7, 8, 8, 9, 9, 10, 11, 
-11, 12, 12, 13, 14, 15, 15, 16, 16, 17, 17, 9, 4, 5, 17, 24, 
-23, 16, 12, 10, 17, 30, 37, 26, 15, 15, 19, 22, 10, 10, 11, 14, 
-16, 19, 19, 9, 2, 1, 2, 4, 6, 25, 24, 28, 28, 26, 23, 22, 
-24, 17, 23, 25, 19, 24, 23, 17, 19, 19, 21, 22, 21, 21, 21, 22, 
-5, 5, 5, 5, 5, 6, 6, 6, 6, 7, 7, 8, 8, 8, 9, 10, 
-10, 11, 11, 12, 12, 13, 14, 14, 15, 15, 15, 6, 4, 6, 21, 26, 
-20, 27, 28, 24, 29, 35, 37, 26, 18, 23, 25, 27, 26, 22, 20, 20, 
-22, 21, 20, 11, 4, 2, 2, 3, 4, 21, 24, 27, 27, 26, 22, 18, 
-20, 18, 22, 23, 18, 23, 23, 19, 19, 19, 21, 21, 21, 21, 21, 22, 
-5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 7, 7, 7, 9, 21, 19, 
-8, 10, 10, 11, 11, 12, 13, 13, 14, 13, 12, 4, 4, 6, 24, 34, 
-35, 32, 29, 30, 35, 36, 35, 24, 21, 23, 29, 29, 27, 29, 30, 29, 
-27, 23, 18, 11, 5, 3, 3, 1, 4, 20, 21, 25, 24, 22, 19, 13, 
-17, 21, 24, 23, 18, 22, 23, 21, 18, 21, 20, 20, 21, 21, 21, 22, 
-5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 7, 7, 7, 18, 38, 30, 
-17, 9, 10, 10, 11, 11, 12, 13, 13, 12, 10, 4, 4, 5, 23, 36, 
-38, 38, 37, 38, 35, 36, 33, 23, 21, 21, 28, 32, 34, 34, 32, 31, 
-28, 22, 17, 10, 6, 4, 4, 2, 5, 21, 20, 23, 24, 25, 23, 14, 
-18, 24, 24, 23, 17, 21, 23, 21, 19, 21, 21, 20, 20, 21, 21, 21, 
-5, 4, 5, 5, 5, 5, 5, 5, 6, 6, 6, 7, 7, 12, 34, 29, 
-22, 10, 9, 8, 7, 7, 8, 8, 10, 8, 5, 4, 7, 7, 19, 37, 
-38, 39, 38, 37, 35, 37, 33, 24, 18, 19, 27, 33, 35, 34, 32, 30, 
-26, 20, 16, 11, 7, 8, 7, 3, 8, 25, 22, 26, 25, 22, 22, 20, 
-19, 22, 24, 24, 19, 22, 22, 21, 20, 20, 20, 20, 20, 21, 21, 21, 
-5, 4, 4, 4, 5, 5, 5, 5, 5, 6, 6, 6, 7, 7, 31, 27, 
-21, 13, 9, 6, 3, 3, 5, 7, 7, 6, 6, 7, 10, 5, 17, 36, 
-37, 38, 37, 34, 33, 38, 35, 25, 20, 18, 24, 32, 33, 33, 30, 27, 
-24, 19, 16, 12, 10, 11, 8, 4, 16, 25, 23, 25, 24, 24, 24, 22, 
-19, 22, 23, 22, 18, 22, 22, 21, 20, 20, 18, 21, 21, 20, 20, 21, 
-5, 4, 4, 4, 5, 5, 5, 5, 5, 6, 6, 6, 7, 7, 31, 27, 
-20, 15, 8, 6, 4, 4, 6, 8, 11, 10, 11, 10, 8, 3, 15, 32, 
-34, 34, 30, 23, 31, 35, 31, 20, 19, 19, 16, 23, 28, 28, 24, 21, 
-19, 18, 16, 12, 13, 13, 6, 11, 23, 25, 22, 25, 25, 24, 22, 21, 
-19, 19, 23, 23, 19, 20, 22, 20, 19, 21, 20, 21, 21, 20, 20, 20, 
-4, 4, 4, 3, 4, 4, 5, 5, 5, 6, 6, 6, 6, 15, 35, 27, 
-19, 12, 7, 6, 3, 4, 5, 6, 8, 8, 9, 8, 5, 2, 7, 27, 
-30, 30, 24, 29, 25, 20, 17, 10, 9, 11, 16, 21, 22, 23, 18, 16, 
-15, 17, 17, 8, 3, 4, 8, 19, 21, 20, 22, 25, 26, 25, 24, 21, 
-22, 19, 22, 22, 22, 19, 22, 19, 19, 21, 21, 21, 20, 20, 20, 20, 
-4, 4, 4, 3, 4, 4, 5, 5, 5, 6, 6, 6, 6, 29, 35, 24, 
-17, 11, 6, 4, 3, 4, 4, 5, 7, 7, 7, 6, 6, 7, 9, 22, 
-30, 28, 23, 31, 30, 22, 16, 11, 11, 14, 17, 18, 19, 20, 20, 16, 
-18, 18, 16, 4, 1, 1, 6, 10, 16, 22, 22, 26, 23, 22, 21, 19, 
-20, 19, 18, 19, 19, 17, 19, 19, 19, 20, 21, 20, 20, 19, 20, 20, 
-4, 4, 4, 2, 4, 4, 5, 5, 5, 5, 5, 6, 11, 38, 33, 25, 
-15, 9, 6, 6, 7, 7, 8, 8, 9, 8, 6, 6, 8, 9, 12, 18, 
-30, 25, 25, 24, 26, 27, 22, 21, 16, 14, 14, 13, 14, 15, 22, 16, 
-21, 17, 14, 1, 2, 8, 10, 16, 21, 22, 22, 23, 18, 24, 22, 23, 
-20, 21, 20, 21, 20, 19, 19, 18, 14, 20, 21, 19, 19, 19, 20, 20, 
-3, 3, 4, 2, 4, 4, 4, 5, 5, 5, 5, 6, 27, 36, 32, 23, 
-13, 7, 6, 6, 8, 9, 9, 10, 10, 10, 9, 7, 7, 8, 7, 10, 
-31, 29, 22, 21, 18, 21, 22, 21, 17, 14, 11, 8, 14, 19, 23, 15, 
-19, 15, 9, 1, 3, 13, 17, 21, 22, 24, 21, 23, 17, 23, 23, 23, 
-21, 21, 21, 22, 21, 20, 18, 17, 14, 19, 21, 18, 19, 19, 19, 20, 
-4, 4, 4, 2, 4, 4, 4, 5, 5, 5, 5, 15, 39, 36, 30, 21, 
-10, 7, 7, 7, 8, 9, 9, 10, 10, 11, 10, 6, 6, 6, 6, 4, 
-26, 29, 26, 30, 24, 23, 22, 19, 18, 15, 15, 17, 20, 22, 16, 17, 
-16, 14, 4, 1, 2, 9, 20, 22, 24, 24, 21, 23, 19, 22, 23, 23, 
-22, 19, 21, 22, 21, 19, 17, 17, 14, 19, 20, 17, 19, 19, 19, 19, 
-3, 4, 4, 2, 4, 4, 4, 5, 5, 5, 9, 37, 38, 35, 30, 24, 
-9, 7, 7, 7, 8, 9, 9, 10, 10, 11, 11, 7, 6, 6, 6, 5, 
-14, 29, 26, 30, 27, 26, 28, 26, 23, 18, 16, 19, 21, 17, 15, 17, 
-14, 11, 1, 1, 3, 4, 14, 21, 22, 22, 19, 23, 19, 22, 24, 23, 
-23, 20, 20, 21, 21, 20, 16, 16, 12, 18, 19, 17, 19, 19, 19, 19, 
-3, 4, 4, 2, 4, 4, 4, 5, 5, 6, 31, 40, 40, 39, 36, 29, 
-17, 11, 7, 7, 8, 8, 8, 8, 6, 7, 11, 8, 7, 4, 4, 6, 
-6, 22, 27, 30, 28, 23, 18, 17, 15, 15, 18, 21, 19, 16, 16, 15, 
-13, 7, 0, 1, 3, 6, 6, 16, 21, 19, 14, 23, 19, 21, 23, 23, 
-22, 22, 19, 20, 21, 20, 16, 16, 12, 17, 19, 16, 19, 19, 19, 19, 
-4, 4, 5, 3, 5, 5, 5, 6, 6, 29, 40, 42, 42, 40, 41, 39, 
-39, 39, 38, 34, 28, 17, 6, 4, 4, 4, 6, 8, 9, 5, 4, 8, 
-7, 8, 24, 29, 32, 32, 31, 29, 26, 25, 25, 21, 16, 12, 13, 12, 
-12, 5, 1, 2, 3, 7, 16, 6, 12, 15, 13, 23, 19, 20, 23, 22, 
-22, 22, 21, 18, 20, 19, 15, 16, 11, 17, 19, 12, 7, 8, 8, 8, 
-5, 5, 6, 4, 7, 7, 7, 8, 18, 39, 40, 41, 37, 37, 38, 39, 
-39, 38, 38, 36, 35, 33, 13, 2, 2, 2, 2, 3, 5, 7, 2, 5, 
-10, 7, 15, 23, 27, 31, 30, 28, 27, 24, 19, 15, 11, 9, 11, 13, 
-12, 5, 1, 2, 3, 6, 10, 5, 14, 15, 19, 22, 20, 19, 22, 22, 
-22, 22, 21, 20, 18, 16, 12, 12, 10, 12, 13, 10, 6, 6, 5, 4, 
-3, 3, 3, 3, 5, 5, 6, 12, 32, 39, 40, 37, 33, 32, 33, 33, 
-34, 35, 35, 33, 33, 32, 22, 8, 7, 4, 4, 5, 5, 5, 15, 18, 
-14, 5, 17, 21, 16, 21, 22, 19, 20, 15, 10, 7, 7, 10, 13, 14, 
-12, 4, 2, 2, 3, 2, 6, 15, 22, 20, 17, 23, 20, 19, 22, 22, 
-21, 21, 21, 20, 17, 3, 2, 3, 3, 4, 4, 4, 5, 5, 6, 6, 
-3, 2, 3, 3, 5, 5, 6, 23, 38, 39, 37, 32, 29, 25, 24, 25, 
-27, 28, 29, 28, 29, 28, 20, 14, 7, 9, 11, 13, 10, 7, 11, 4, 
-4, 3, 17, 28, 19, 9, 8, 7, 8, 7, 6, 8, 10, 14, 15, 15, 
-11, 2, 2, 2, 2, 2, 5, 15, 22, 20, 18, 22, 20, 19, 21, 22, 
-21, 21, 21, 20, 18, 3, 3, 3, 3, 3, 3, 2, 2, 2, 2, 2, 
-18, 9, 7, 4, 4, 5, 16, 37, 38, 39, 40, 39, 35, 30, 21, 14, 
-17, 21, 21, 21, 21, 15, 14, 10, 7, 9, 7, 7, 6, 4, 6, 4, 
-4, 5, 17, 31, 28, 22, 16, 14, 13, 12, 12, 12, 14, 16, 17, 17, 
-12, 2, 1, 1, 1, 1, 3, 3, 6, 14, 19, 22, 18, 18, 21, 21, 
-21, 21, 21, 20, 18, 3, 4, 6, 7, 20, 13, 6, 5, 4, 3, 2, 
-5, 2, 8, 12, 12, 17, 36, 38, 38, 39, 42, 41, 37, 35, 35, 35, 
-37, 37, 35, 33, 28, 22, 13, 7, 7, 6, 6, 5, 5, 4, 4, 4, 
-5, 5, 16, 31, 31, 27, 25, 24, 21, 17, 16, 16, 17, 18, 19, 18, 
-12, 2, 1, 2, 2, 2, 3, 4, 4, 4, 4, 6, 6, 9, 15, 20, 
-21, 19, 18, 19, 17, 3, 4, 5, 7, 32, 28, 8, 8, 7, 7, 3, 
-2, 2, 2, 4, 14, 34, 37, 38, 37, 40, 42, 40, 35, 34, 35, 35, 
-35, 35, 34, 33, 31, 30, 24, 7, 7, 4, 5, 5, 5, 4, 3, 5, 
-5, 5, 8, 30, 32, 29, 27, 26, 24, 18, 17, 18, 19, 20, 20, 18, 
-6, 1, 2, 3, 2, 3, 4, 4, 3, 3, 5, 5, 5, 5, 6, 8, 
-8, 3, 3, 8, 16, 3, 3, 4, 5, 18, 21, 7, 6, 6, 6, 3, 
-2, 2, 2, 10, 33, 37, 37, 37, 36, 37, 37, 31, 24, 22, 23, 23, 
-25, 27, 28, 27, 26, 25, 21, 7, 6, 4, 5, 5, 5, 4, 3, 6, 
-5, 5, 5, 19, 30, 30, 29, 25, 21, 20, 20, 22, 23, 22, 20, 9, 
-2, 2, 3, 4, 3, 3, 4, 4, 2, 5, 5, 5, 6, 5, 6, 5, 
-4, 6, 7, 8, 8, 5, 4, 4, 5, 10, 10, 6, 5, 5, 4, 3, 
-1, 2, 14, 35, 37, 37, 36, 36, 36, 34, 25, 19, 18, 18, 18, 19, 
-21, 21, 22, 21, 20, 17, 11, 6, 4, 4, 5, 5, 5, 4, 4, 6, 
-5, 5, 4, 4, 19, 29, 30, 29, 27, 25, 25, 25, 25, 20, 7, 2, 
-2, 3, 3, 4, 3, 3, 4, 2, 4, 5, 6, 6, 6, 6, 6, 5, 
-6, 6, 7, 7, 7, 7, 7, 6, 4, 7, 8, 5, 5, 5, 4, 3, 
-2, 16, 35, 36, 36, 36, 35, 36, 37, 38, 35, 29, 25, 22, 19, 19, 
-25, 24, 21, 18, 14, 8, 6, 6, 3, 5, 5, 5, 6, 2, 5, 6, 
-6, 5, 5, 4, 3, 7, 16, 22, 24, 24, 21, 16, 9, 3, 2, 2, 
-3, 3, 4, 3, 3, 4, 4, 3, 5, 6, 5, 6, 6, 6, 5, 6, 
-6, 6, 7, 6, 7, 7, 7, 7, 7, 12, 14, 6, 5, 4, 4, 3, 
-20, 34, 35, 36, 35, 34, 34, 35, 37, 40, 37, 32, 30, 31, 32, 32, 
-33, 32, 30, 28, 24, 9, 5, 6, 2, 5, 5, 6, 5, 3, 4, 6, 
-5, 5, 5, 4, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 4, 
-4, 3, 4, 2, 4, 4, 3, 4, 6, 6, 6, 6, 6, 6, 5, 6, 
-6, 6, 6, 6, 6, 6, 6, 6, 6, 7, 6, 4, 4, 4, 4, 2, 
-34, 34, 34, 34, 33, 33, 34, 34, 35, 35, 29, 26, 23, 24, 25, 26, 
-28, 26, 25, 24, 21, 8, 4, 6, 2, 4, 6, 6, 3, 4, 4, 5, 
-5, 5, 4, 4, 4, 4, 4, 4, 3, 3, 3, 3, 3, 4, 4, 4, 
-4, 4, 4, 3, 4, 5, 3, 5, 6, 5, 6, 6, 6, 6, 5, 6, 
-6, 6, 5, 5, 6, 6, 6, 6, 5, 4, 4, 5, 4, 4, 4, 2, 
-34, 34, 33, 33, 32, 32, 33, 34, 32, 32, 26, 19, 19, 19, 20, 20, 
-20, 19, 18, 16, 10, 5, 4, 5, 2, 5, 5, 5, 2, 4, 5, 5, 
-5, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 
-4, 5, 3, 4, 4, 5, 4, 6, 6, 6, 6, 5, 6, 5, 5, 6, 
-6, 5, 4, 6, 6, 6, 6, 5, 5, 6, 6, 6, 4, 4, 4, 2, 
-34, 33, 33, 35, 32, 32, 34, 36, 35, 33, 21, 17, 14, 13, 15, 18, 
-16, 13, 8, 6, 4, 4, 4, 4, 3, 4, 4, 4, 3, 3, 4, 4, 
-5, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 4, 
-4, 5, 3, 5, 5, 4, 4, 6, 6, 6, 6, 6, 6, 5, 5, 6, 
-6, 4, 6, 6, 6, 6, 5, 6, 6, 6, 6, 6, 6, 5, 3, 2, 
-35, 34, 33, 33, 31, 26, 33, 36, 38, 35, 28, 24, 25, 29, 30, 29, 
-26, 23, 10, 4, 4, 4, 4, 4, 3, 4, 4, 4, 4, 3, 4, 4, 
-4, 4, 4, 4, 4, 4, 4, 5, 4, 5, 4, 4, 4, 5, 5, 4, 
-5, 5, 3, 5, 6, 5, 4, 6, 6, 6, 6, 6, 6, 5, 5, 6, 
-4, 5, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 5, 4, 1, 1, 
-35, 34, 34, 29, 26, 19, 25, 34, 37, 34, 28, 26, 26, 27, 27, 24, 
-21, 18, 7, 4, 4, 4, 4, 4, 2, 4, 4, 3, 4, 3, 4, 4, 
-4, 4, 4, 4, 4, 4, 4, 4, 5, 4, 4, 4, 4, 5, 5, 4, 
-5, 5, 3, 5, 5, 6, 4, 6, 6, 6, 5, 5, 6, 5, 5, 5, 
-4, 5, 6, 6, 6, 6, 6, 6, 6, 6, 6, 5, 4, 6, 4, 1, 
-33, 30, 25, 20, 18, 15, 14, 27, 29, 26, 23, 20, 21, 19, 18, 15, 
-11, 6, 3, 3, 4, 4, 4, 3, 2, 4, 4, 4, 4, 3, 4, 4, 
-4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 
-4, 5, 3, 4, 5, 5, 5, 5, 5, 6, 4, 5, 6, 5, 4, 4, 
-4, 6, 6, 6, 6, 6, 6, 6, 5, 6, 6, 6, 6, 6, 6, 5, 
-29, 26, 25, 24, 20, 16, 12, 10, 16, 15, 14, 14, 13, 11, 8, 5, 
-3, 2, 3, 3, 3, 3, 4, 3, 2, 3, 3, 4, 4, 2, 3, 4, 
-4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 4, 4, 5, 4, 5, 
-4, 4, 4, 4, 5, 5, 5, 5, 5, 6, 4, 5, 5, 5, 4, 4, 
-5, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 
-30, 30, 28, 25, 19, 12, 7, 5, 4, 5, 8, 8, 5, 3, 3, 3, 
-3, 2, 3, 3, 3, 3, 3, 3, 2, 3, 3, 3, 3, 2, 3, 3, 
-4, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 5, 4, 
-3, 4, 4, 4, 5, 4, 5, 5, 5, 6, 3, 4, 5, 4, 3, 4, 
-6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 8, 14, 20, 24, 25, 24, 
-32, 29, 25, 19, 12, 7, 4, 3, 3, 3, 3, 3, 3, 2, 2, 2, 
-2, 2, 3, 3, 3, 3, 3, 3, 2, 3, 3, 3, 3, 2, 3, 3, 
-4, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 4, 
-3, 4, 4, 4, 5, 4, 4, 5, 5, 5, 3, 4, 4, 4, 3, 5, 
-5, 6, 6, 6, 6, 6, 6, 6, 8, 20, 29, 30, 29, 29, 28, 27, 
-28, 24, 19, 12, 7, 3, 3, 3, 2, 2, 2, 2, 2, 2, 2, 1, 
-1, 1, 2, 2, 2, 2, 2, 2, 1, 2, 3, 3, 3, 2, 3, 3, 
-3, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 3, 
-4, 4, 4, 4, 5, 4, 4, 4, 5, 5, 3, 4, 4, 4, 2, 5, 
-4, 6, 6, 6, 6, 6, 6, 12, 29, 32, 32, 31, 31, 30, 30, 30, 
-23, 18, 12, 8, 3, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, 
-1, 1, 2, 2, 2, 2, 2, 2, 1, 2, 3, 3, 3, 2, 3, 3, 
-3, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 4, 4, 4, 3, 3, 
-4, 4, 4, 4, 4, 3, 4, 4, 5, 4, 3, 4, 4, 4, 2, 5, 
-3, 5, 6, 5, 6, 6, 11, 30, 32, 32, 32, 33, 32, 32, 31, 30, 
-16, 12, 8, 3, 1, 1, 1, 1, 2, 2, 2, 2, 1, 1, 1, 1, 
-1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 2, 3, 3, 
-3, 3, 3, 3, 3, 3, 4, 3, 4, 4, 4, 4, 4, 4, 3, 4, 
-4, 4, 4, 3, 4, 3, 4, 4, 4, 4, 3, 3, 4, 4, 2, 4, 
-2, 5, 5, 5, 6, 7, 28, 32, 32, 33, 33, 33, 33, 32, 32, 32, 
-11, 8, 6, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
-1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 2, 3, 3, 
-3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 
-4, 4, 4, 4, 4, 3, 4, 4, 4, 4, 3, 3, 4, 3, 3, 4, 
-2, 4, 4, 5, 6, 17, 30, 31, 32, 33, 33, 33, 33, 33, 33, 33, };
-};
-
