+++ CompositeVideo/Color.cpp
@@ -1,0 +1,5 @@
+#include <Arduino.h>
+#include "Color.h"
+
+char Color::hue = 0x00;
+
